requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
decorator==5.1.0
Flask==2.0.2
future==0.18.2
geographiclib==1.52
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
python-dotenv==0.19.1
ratelim==0.1.6
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.2
